refactor(memory): isolate langchaingo shim into langchainadapter subpackage
Move the langchain-compatible memory adapter into a new
memory/langchainadapter subpackage, making it the sole importer of
github.com/tmc/langchaingo in the memory tree. The core memory package
now has zero langchaingo dependencies (verified via go list -deps).

- git mv langchain_adapter.go -> langchainadapter/adapter.go
- delete zero-ref constructors NewLangChainMemory and
  NewConversationTokenBufferMemory
- repoint memory_basic / memory_chatbot examples to the subpackage
- fix stale prefixes and drop ConversationTokenBuffer section in
  the EN/CN LangChain memory integration docs

Closes #100
