+++ docs/issue#0100.html
@@ -1,0 +1,99 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Implementation Notes — Issue #0100</title>
+  <style>
+    * { margin: 0; padding: 0; box-sizing: border-box; }
+    body {
+      font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
+      background: #FAF9F6;
+      color: #1a1a1a;
+      padding: 2.5rem 2rem;
+      line-height: 1.7;
+    }
+    .container { max-width: 800px; margin: 0 auto; }
+    h1 { font-size: 1.375rem; font-weight: 700; margin-bottom: 0.25rem; }
+    .meta { color: #8B8680; font-size: 0.8125rem; margin-bottom: 2rem; }
+    h2 {
+      font-size: 1rem;
+      font-weight: 600;
+      margin-top: 2rem;
+      margin-bottom: 0.75rem;
+      padding-bottom: 0.375rem;
+      border-bottom: 1px solid #E8E4DE;
+      display: flex; align-items: center; gap: 0.5rem;
+    }
+    .dot { width: 8px; height: 8px; border-radius: 50%; display: inline-block; }
+    .dot.design { background: #5B8A72; }
+    .dot.deviation { background: #D97757; }
+    .dot.tradeoff { background: #4A6FA5; }
+    .dot.question { background: #D4A843; }
+    .item {
+      background: #FFFFFF;
+      border: 1px solid #E8E4DE;
+      border-radius: 8px;
+      padding: 1rem 1.25rem;
+      margin-bottom: 0.75rem;
+      box-shadow: 0 1px 3px rgba(0,0,0,0.03);
+    }
+    .item h3 { font-size: 0.875rem; font-weight: 600; margin-bottom: 0.375rem; }
+    .item p { font-size: 0.8125rem; color: #4A4540; margin-bottom: 0.375rem; }
+    .label {
+      display: inline-block;
+      font-size: 0.6875rem;
+      font-weight: 500;
+      padding: 0.125rem 0.5rem;
+      border-radius: 4px;
+      margin-right: 0.375rem;
+    }
+    .label-design { background: #F5F8F6; color: #5B8A72; border: 1px solid #5B8A72; }
+    .label-deviation { background: #FFF5F0; color: #D97757; border: 1px solid #D97757; }
+    .label-tradeoff { background: #F0F4F8; color: #4A6FA5; border: 1px solid #4A6FA5; }
+    .label-question { background: #FDF8F0; color: #D4A843; border: 1px solid #D4A843; }
+    .none { color: #B0AAA4; font-style: italic; font-size: 0.8125rem; }
+    .footer { text-align: center; margin-top: 2.5rem; color: #B0AAA4; font-size: 0.6875rem; }
+    code { background: #F0EEE9; padding: 0.05rem 0.3rem; border-radius: 3px; font-size: 0.78rem; }
+  </style>
+</head>
+<body>
+  <div class="container">
+    <h1>Implementation Notes</h1>
+    <p class="meta">Issue <a href="https://github.com/smallnest/langgraphgo/issues/100">#100</a> &mdash; 精简: 移除/隔离 memory 中的 langchaingo 兼容垫片 &mdash; 2026-07-08</p>
+
+    <h2><span class="dot design"></span> Design Decisions</h2>
+    <div class="item">
+      <h3><span class="label label-design">Decision</span> New subpackage <code>memory/langchainadapter</code> as the sole langchaingo importer</h3>
+      <p><strong>Ambiguity:</strong> The issue said "迁移到独立子包(如 memory/langchainadapter)" but did not name the package or specify what moves.</p>
+      <p><strong>Choice:</strong> Moved the entire adapter file (types <code>ChatMemory</code>, <code>LangChainMemory</code>, <code>ChatMessageHistory</code> and their methods) verbatim into <code>package langchainadapter</code> via <code>git mv</code>, preserving history.</p>
+      <p><strong>Rationale:</strong> <code>ChatMemory</code> is referenced only by the adapter itself, so nothing had to stay in the core package. <code>go list -deps ./memory</code> now reports zero <code>tmc/langchaingo</code> dependencies — the stated goal.</p>
+    </div>
+
+    <h2><span class="dot deviation"></span> Deviations</h2>
+    <div class="item">
+      <h3><span class="label label-deviation">Deviation</span> Examples repointed to the adapter, not rewritten to native Buffer/SlidingWindow</h3>
+      <p><strong>Spec:</strong> Suggestion #3 said "更新 examples 使用原生 Buffer/SlidingWindow 构造函数".</p>
+      <p><strong>Implemented:</strong> <code>examples/memory_basic</code> and <code>examples/memory_chatbot</code> keep using the langchain adapter, only switching their import to the new <code>langchainadapter</code> subpackage.</p>
+      <p><strong>Why:</strong> The native <code>BufferMemory</code> API is structurally different (message-object <code>AddMessage</code>/<code>GetContext</code> vs the adapter's map-based <code>SaveContext</code>/<code>LoadMemoryVariables</code> and langchain options like <code>WithMemoryKey</code>/<code>WithReturnMessages</code>). These examples exist specifically to demonstrate the langchain-compatible surface, so rewriting them would delete that demonstration rather than update it. The dependency-isolation goal is fully met regardless of which constructor the examples call.</p>
+    </div>
+
+    <h2><span class="dot tradeoff"></span> Tradeoffs</h2>
+    <div class="item">
+      <h3><span class="label label-tradeoff">Tradeoff</span> Fixed the affected docs in place instead of a full rewrite</h3>
+      <p><strong>Alternatives:</strong> (a) leave docs untouched; (b) fully rewrite the LangChain integration docs; (c) minimally correct only what this change broke.</p>
+      <p><strong>Chosen (c):</strong> Removed the <code>ConversationTokenBuffer</code> section (its constructor was deleted), corrected the long-stale <code>prebuilt.</code> prefix to <code>langchainadapter.</code>, and added an import snippet — in both EN and CN docs.</p>
+      <p><strong>Why:</strong> (a) would ship docs referencing a deleted symbol; (b) is a separate docs task beyond this issue's scope. (c) keeps the shipped tree self-consistent with minimal blast radius.</p>
+    </div>
+
+    <h2><span class="dot question"></span> Open Questions</h2>
+    <div class="item">
+      <h3><span class="label label-question">Question</span> Release-note the API break</h3>
+      <p><strong>Assumption:</strong> Deleting <code>NewLangChainMemory</code> / <code>NewConversationTokenBufferMemory</code> (both zero internal references) and relocating the surviving constructors to <code>memory/langchainadapter</code> is an acceptable break.</p>
+      <p><strong>Verify:</strong> External callers importing these from <code>memory</code> must update their import path; the token-buffer helper has no drop-in replacement in this repo. The issue itself flagged this needs a release note.</p>
+    </div>
+
+    <p class="footer">Generated by goal-workflow /note-it</p>
+  </div>
+</body>
+</html>
