+++ examples/memory_chatbot/main.go
@@ -6,7 +6,7 @@ import (
 	"log"
 
 	"github.com/smallnest/langgraphgo/llmtypes"
-	"github.com/smallnest/langgraphgo/memory"
+	langchainadapter "github.com/smallnest/langgraphgo/memory/langchainadapter"
 	langchainmemory "github.com/tmc/langchaingo/memory"
 )
 
@@ -31,18 +31,18 @@ func main() {
 // runChatbotSimulation simulates a chatbot conversation with memory
 func runChatbotSimulation(ctx context.Context, memoryType string, windowSize int) {
 	// Create memory based on type
-	var mem *memory.LangChainMemory
+	var mem *langchainadapter.LangChainMemory
 	switch memoryType {
 	case "buffer":
-		mem = memory.NewConversationBufferMemory(
+		mem = langchainadapter.NewConversationBufferMemory(
 			langchainmemory.WithReturnMessages(true),
 		)
 	case "window":
-		mem = memory.NewConversationWindowBufferMemory(windowSize,
+		mem = langchainadapter.NewConversationWindowBufferMemory(windowSize,
 			langchainmemory.WithReturnMessages(true),
 		)
 	default:
-		mem = memory.NewConversationBufferMemory(
+		mem = langchainadapter.NewConversationBufferMemory(
 			langchainmemory.WithReturnMessages(true),
 		)
 	}
@@ -124,14 +124,14 @@ func runChatbotSimulation(ctx context.Context, memoryType string, windowSize int
 // demonstrateMemoryPersistence shows how memory persists across interactions
 func demonstrateMemoryPersistence(ctx context.Context) {
 	// Create a custom chat history
-	chatHistory := memory.NewChatMessageHistory()
+	chatHistory := langchainadapter.NewChatMessageHistory()
 
 	// Add initial messages
 	chatHistory.AddUserMessage(ctx, "I'm learning about LangGraph")
 	chatHistory.AddAIMessage(ctx, "That's great! LangGraph is a powerful framework for building stateful, multi-actor applications.")
 
 	// Create memory with the custom chat history
-	mem := memory.NewConversationBufferMemory(
+	mem := langchainadapter.NewConversationBufferMemory(
 		langchainmemory.WithChatHistory(chatHistory.GetHistory()),
 		langchainmemory.WithReturnMessages(true),
 	)
