+++ memory/langchainadapter/adapter_test.go
@@ -1,4 +1,4 @@
-package memory
+package langchainadapter
 
 import (
 	"context"
